Initial commit
Added all SQL files I previously had stored in another repository. They will change item templates and creature loot templates to make as many quest items as possible be looted by all members of a group. Items that are normally looted from objects are added to creatures in the same area as the objects when possible. Objects that lack viable creatures near them are simply set to a lower respawn time.
Added scripts that change how certain triggers are handled to make sure they give credit to all members of a group.

I'm not going to go into detail on what is added in this commit, check the code if you're interested. Commits after this one will include details, though.