Major rewrite
I decided to rewrite the SQL scripts from the ground up. This time I intend to properly document what is changed and spend some time researching new entries I add myself to make them seem fair.

There are a few items that you need from creatures that are obviously not changed by this. I will do my research for those items to see if they should be changed or if there's a reliable way of obtaining them as a group without having to change their flag.